test: replaced __mocks__ with inline while fixing the react-multi-comp rule of eslint
